Cleaned extension classes
